+++ tests/test_graph.py
@@ -155,37 +155,3 @@ def test_build_graph_node_features_unsupported() -> None:
 
     with pytest.raises(ValueError, match="Unsupported property type:"):
         _ = graph.graph_from_adjacency_matrix(A, property)
-
-
-@pytest.mark.skipif(
-    # Run test if all supported backends are installed
-    not set(spyrmsd.graph._supported_backends) <= set(spyrmsd.available_backends),
-    reason="Not all of the required backends are installed",
-)
-def test_set_backend() -> None:
-    import graph_tool as gt
-    import networkx as nx
-    import rustworkx as rx
-
-    A = np.array([[0, 1, 1], [1, 0, 0], [1, 0, 1]])
-
-    spyrmsd.set_backend("networkx")
-    assert spyrmsd.get_backend() == "networkx"
-
-    Gnx = graph.graph_from_adjacency_matrix(A)
-    assert isinstance(Gnx, nx.Graph)
-
-    spyrmsd.set_backend("graph-tool")
-    assert spyrmsd.get_backend() == "graph_tool"
-
-    Ggt = graph.graph_from_adjacency_matrix(A)
-    assert isinstance(Ggt, gt.Graph)
-
-    spyrmsd.set_backend("rustworkx")
-    assert spyrmsd.get_backend() == "rustworkx"
-
-    Grx = graph.graph_from_adjacency_matrix(A)
-    assert isinstance(Grx, rx.PyGraph)
-
-    with pytest.raises(ValueError, match="backend is not recognized or supported"):
-        spyrmsd.set_backend("unknown")
